Reorganized things, added more detailed documentation via README
